third_party/WebKit/LayoutTests/platform/win/virtual/service-worker-navigation-preload/http/tests/serviceworker/chromium/usecounter-on-reload-expected.txt

This is a testharness.js-based test.
PASS usecounter is counted upon reload after registration 
Harness: the test ran to completion.
